feat: change structure
