+++ src/components/index.ts
@@ -4,7 +4,10 @@ export { BaseButton } from "./base-button/index.tsx";
 export { FooterWrapper } from "./common/wrappers/footer-wrapper.tsx";
 export { HeaderWrapper } from "./common/wrappers/header-wrapper.tsx";
 export { CustomFormGroup } from "./custom-form-group/index.tsx";
+export { FilterButton } from "./filter-button/index.tsx";
+export { Newsletter } from "./newsletter/index.tsx";
 export { PrivateRoute } from "./private-route/index.tsx";
 export { PublicRoute } from "./public-route/index.tsx";
 export { RouterProvider } from "./router-provider/index.tsx";
+export { TopInfoSection } from "./top-info-section/index.tsx";
 export { Provider as StoreProvider } from "react-redux";
