+++ src/pages/home/components/category-item/index.tsx
@@ -3,7 +3,7 @@ import React, { useCallback } from "react";
 
 import { colors, fontSizes } from "~/libs/constants/index.ts";
 
-import { type Category } from "../types/index.ts";
+import { type Category } from "../../types/index.ts";
 import {
 	RoundImage,
 	StyledCategoryContainer,
