+++ src/pages/home/components/index.ts
@@ -1,5 +1,4 @@
-export { CategoryCarousel } from "./category-carousel.tsx";
-export { CategoryItem } from "./category-item.tsx";
-export { Newsletter } from "./newsletter.tsx";
-export { Products } from "./products.tsx";
-export { TopInfoSection } from "./top-info-section.tsx";
+export { CategoryCarousel } from "./category-carousel/index.tsx";
+export { CategoryItem } from "./category-item/index.tsx";
+export { ProductCard } from "./product-card/index.tsx";
+export { Products } from "./products/index.tsx";
