+++ src/pages/home/components/product-card/index.tsx
@@ -1,5 +1,6 @@
 import { Box, Typography } from "@mui/material";
-import React from "react";
+import React, { useCallback } from "react";
+import { useNavigate } from "react-router-dom";
 
 import { fontSizes } from "~/libs/constants/font.ts";
 import { type Product } from "~/libs/types/products.ts";
@@ -11,20 +12,27 @@ type ProductCardProperties = {
 };
 
 const ProductCard: React.FC<ProductCardProperties> = ({ product }) => {
+	const navigate = useNavigate();
+
+	const handleClick = useCallback(() => {
+		navigate(`/product/${product.id}`);
+	}, [navigate, product]);
+
 	return (
 		<Box
+			onClick={handleClick}
 			sx={{
 				display: "flex",
 				flexDirection: "column",
 				gap: "12px",
+				height: "360px",
 				maxWidth: "300px",
 			}}
 		>
 			<img
 				alt={product.name}
-				height="322px"
 				src={product.imageUrl}
-				width="100%"
+				style={{ height: "80%", width: "100%" }}
 			/>
 			<StyledProductCardContent>
 				<Typography color="primary" variant="playfairDisplayBold">
