+++ src/pages/home/components/product-card/styles.ts
@@ -1,0 +1,10 @@
+import { Box, styled } from "@mui/material";
+
+const StyledProductCardContent = styled(Box)`
+	align-item: flex-start;
+	display: flex;
+	justify-content: space-between;
+	flex-direction: column;
+`;
+
+export { StyledProductCardContent };
