+++ src/pages/home/components/products/styles.ts
@@ -1,0 +1,12 @@
+import { Box, styled } from "@mui/material";
+
+const StyledProductsContainer = styled(Box)`
+	align-item: flex-start;
+	display: flex;
+	justify-content: space-evenly;
+	gap: 30px;
+	flex-wrap: wrap;
+	padding: 25px 0px;
+`;
+
+export { StyledProductsContainer };
